+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha53] - 2021-09-29
+### Added
+ - Added `nop` URL parameters to `Text` factory.
+ - The `nop` transformation will be default text transformation, if there is no other transformation given.
+
 ## [1.0.0-alpha52] - 2021-06-14
 ### Added
  - Added `frame` and `sec` URL parameters to `Transformation` factory.